Fix function docs for OTP 22+ (#72)
Prior to this commit, erldocs was failing to create an index entry or
page documentation for functions when run against OTP 22. It turns out
that an additional tuple `{since, List}` had been added along with the
function name and arity. This tuple was handled neither in the index or
the page documentation generation. This commit adds two additional
clauses to handle these `since` tuples; however does nothing to display
any `since` information.

There is an existing open ticket regarding displaying `since`
information: #50.
Fixes #70.